Cleanup and fix some tests.
