+++ isaac_arena/assets/object_base.py
@@ -16,16 +16,10 @@ from abc import ABC, abstractmethod
 from enum import Enum
 from typing import Any
 
-from isaaclab.assets import RigidObjectCfg
-from isaaclab.assets.articulation.articulation_cfg import ArticulationCfg
+from isaaclab.assets import ArticulationCfg, AssetBaseCfg, RigidObjectCfg
 from isaaclab.sensors.contact_sensor.contact_sensor_cfg import ContactSensorCfg
 
 from isaac_arena.assets.asset import Asset
-
-# from isaac_arena.assets.object_base import ObjectType
-from isaac_arena.geometry.pose import Pose
-
-# from isaac_arena.assets.object_utils import detect_object_type
 
 
 class ObjectType(Enum):
@@ -53,18 +47,6 @@ class ObjectBase(Asset, ABC):
 
     def get_prim_path(self) -> str:
         return self.prim_path
-
-    # MOVE
-    def set_initial_pose(self, pose: Pose) -> None:
-        self.initial_pose = pose
-
-    # MOVE
-    def get_initial_pose(self) -> Pose | None:
-        return self.initial_pose
-
-    # MOVE
-    def is_initial_pose_set(self) -> bool:
-        return self.initial_pose is not None
 
     def get_cfgs(self) -> dict[str, Any]:
         if self.object_type == ObjectType.RIGID:
@@ -95,5 +77,10 @@ class ObjectBase(Asset, ABC):
 
     @abstractmethod
     def _generate_articulation_cfg(self) -> ArticulationCfg:
+        # Subclasses must implement this method
+        pass
+
+    @abstractmethod
+    def _generate_base_cfg(self) -> AssetBaseCfg:
         # Subclasses must implement this method
         pass
